eureka/api-core/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyjwt==2.8.0

# Validation & Serialization
email-validator==2.1.0
python-dateutil==2.8.2

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.0.0

# Linting & Formatting
ruff==0.1.13
mypy==1.8.0

# Monitoring & Logging
python-json-logger==2.0.7
sentry-sdk==1.39.2

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3

# CORS
fastapi-cors==0.0.6

# AWS/S3 (MinIO compatible)
boto3==1.34.17
botocore==1.34.17

# Background Tasks
celery[redis]==5.3.6

# API Documentation
pyyaml==6.0.1

# Additional dependencies for Session 2
python-jose[cryptography]==3.3.0
alembic==1.13.1
pydantic-settings==2.6.1
